CLOUDP-302068: Simplify abstraction over controller setup (#4124)
Current abstraction makes it hard to parametrise the controllers.
Parametrisation is important when it comes to merging the enterprise and
community operators where we would need to reduce reliance on
environment arguments deep in the call stack to be able to run
controllers of both operators in the same manager.