CLOUDP-288551: Block scaling both ways at runtime (#4099)
# Summary

This PR blocks reconciliation in case we attempt to scale components up
and down at the same time (both ways).

When changing member distribution of a component, they should only scale
down, or up.
A component can be a shard, config servers or mongos
E.g (for a distribution of a component over 3 clusters):
- Scaling from 1 2 2 to 3 3 2 would be valid
- Scaling from 2 2 2 to 1 2 3 would be invalid, as the component is
scaling from 1 to 2 members on cluster 1, but from 2 to 3 on cluster 3.

Multiple cases are illustrated in unit tests.

We need to be defensive in these cases, because it can create
inconsistencies in the behaviour of the operator. For example when we
decide if we should publish the automation config first, we check at the
cluster level, so we can detect that we scale down, whereas the global
replica set (across all clusters) is scaling up.

This change will also impact single cluster sharded clusters (e.g
scaling up shards, but scaling down config servers)

**Note:** we chose to do this check at runtime rather than in webhooks.
Doing it at the webhook level would require to duplicate a lot of logic
from the reconciler, to compute the desired configuration based on shard
overrides, member config...
With the scalers available, it is much easier and makes more sense to do
it at runtime.

## List of changes

- New method in sharded reconciler to block reconciliation if needed.
- Fixed a bug which might cause losing the current sizes from the
deployment state when upgrading the operator from <=1.27
(https://github.com/10gen/ops-manager-kubernetes/pull/4099/commits/0fa2d50acdfa314d69a4579fbe58bd29d8079f32)
- Unit test for this mechanism in various scaling scenarios.
- Modified the `MultiClusterReplicaSetScaler` interface, to avoid type
casting.
- Fixed E2E tests that were attempting both ways reconciliation.
- Added a constant for "slaney", the default name of a sharded cluster
in unit tests.

## Documentation

This ticket will be closed with documentation changes required to update
our public documentation.

## Proof of Work

Unit test `TestBlockReconcileScalingBothWays`
Example error message:
```
Cannot perform scale up and scale down operations at the same time. Scaling Up: [Component=shard idx 0, Cluster=__default, Current=0, Desired=2;], Scaling Down: [Component=configSrv, Cluster=__default, Current=3, Desired=2;]"
```

**Note:** The e2e test `e2e_operator_upgrade_sharded_cluster` is failing
and we are aware of that. It is related to how the state configmap is
handled during a migration. We opened this PR nonetheless as this
failure is not directly related to these changes, but we are working on
adding a fix.

---------

Co-authored-by: Łukasz Sierant <lukasz.sierant@mongodb.com>