CLOUDP-294373: Configure ClusterIP services for mongos (#4134)
# Summary

When configuring a multi-cluster sharded deployment with external access
but no external domain, the operator was not creating clusterIP services
for the components. They could not reach each other.

## Proof of Work

New test ensuring we can deploy a multi-cluster sharded resource with
`externalAccess` defined, but no `externalDomain`, with a service mesh.
The test was failing without the fix, as pods couldn't communicate.