CLOUDP-302068: Refactor `GetMongoVersionForAutomationConfig` (#4140)
Refactors `GetMongoVersionForAutomationConfig` in a way that it no
longer reads `MDB_ASSUME_ENTERPRISE_IMAGE` and `MONGODB_IMAGE` env vars
directly and receives these in arguments.

This will help with merging the community operator and the enterprise
operators into a single process.

Note that more refactoring is needed to be able to merge these. There
are several places with `FIXME(Mikalai)` comments and a follow up for
these is cumming in the next iteration. The effort was split into
multiple parts for more manageable code review.