Moving common MCO API structs to separate package (#4211)
This PR moves some common structures MCO and MDB depend on to a separate
common package to allow to break some circular dependencies for search.

This is an intermediary step, for the final MCK codebase the structures
will be flattened and refactored even more. But this allows to start
working on Search project straight away.

The change doesn't change CRDs. The MongoDBCommunity.spec.search field
is only temporarily and will be soon removed. It helps bootstraping the
development without having the proper MongoDBSearch CRD around.

Additionally, `ST1016` stylecheck was disabled as it suddenly started
reporting issued in the newly introduced package, despite being it the
case for other api packages.