CLOUDP-290261 [Ops Manager] When irsaEnabled == true, s3SecretRef should not be required (#4226)
# Summary

To use IRSA for OpsManager resource you have to set
`spec.backup.s3Stores.irsaEnabled `as true. The `s3SecretRef` is still
required parameter, but by IRSA feature definition it is not needed and
not used in the code.

DOCSP ticket with docs changes ->
https://jira.mongodb.org/browse/DOCSP-46411

## Proof of Work

Added new unit tests to verify optionality for `s3SecretRef` when AWS
IRSA is enabled

## Checklist
- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [x] Have you checked for release_note changes?

## Reminder (Please remove this when merging)
- Please try to Approve or Reject Changes the PR, keep PRs in review as
short as possible
- Our Short Guide for PRs:
[Link](REDACTED)
- Remember the following Communication Standards - use comment prefixes
for clarity:
  * **blocking**: Must be addressed before approval.
  * **follow-up**: Can be addressed in a later PR or ticket.
  * **q**: Clarifying question.
  * **nit**: Non-blocking suggestions.
  * **note**: Side-note, non-actionable. Example: Praise
  * --> no prefix is considered a question