CLOUDP-303641 , CLOUDP-303642 - Fix appdb mc hostnames (#4200)
# Summary

Hostnames for AppDB were set up incorrectly (using the headless services
instead of pod services) for monitoring in Ops Manager. This happened
for 2 reasons:
1. We set up monitoring for AppDB with the headless hostnames by calling
the "/hosts" endpoint of OM.
2. Even after fixing point 1, OM would still choose the headless
hostnames (due to having more dots ".")
More details about the underlying issue can be found in the TD
https://docs.google.com/document/d/1fMWov8gbWWV9Th2LkiHIQ9dDBYr3yAX4UEF7svECIXQ/edit?tab=t.0#heading=h.my8l8kwcujpo

Point 1 was fixed in `dns.go`.
Point 2 was fixed by calling the agent endpoint
"/group/v2/addPreferredHostname". This is not ideal, since the endpoint
is not authenticated using our regular OM credentials, but using the
agent api key that the operator generates and stores in a secret.

As mentioned in the TD, we have created a ticket on the EA side to make
sure these endpoints don't get deleted and also requested a public or
private endpoint we can use instead:
https://jira.mongodb.org/browse/CLOUDP-308115

## Proof of Work

E2E tests are now modified to check:
1. Hostnames are correct - by calling "/hosts" endpoint
2. Preferred hostnames are set - by calling "/group/v2/info"
3. Monitoring works by checking that there is monitoring data present
(this was already present but it would always pass due to a bug)

![image](https://github.com/user-attachments/assets/6aa33bf2-8acf-4020-a6b6-6f70d564b3b1)

![image](https://github.com/user-attachments/assets/a69d7f54-09b4-4f5e-a22f-52a13a253ada)

![image](https://github.com/user-attachments/assets/685104e8-af6b-4c59-ba62-a7de241ae97c)

## Checklist
- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [ ] Have you checked for release_note changes?

## Reminder (Please remove this when merging)
- Please try to Approve or Reject Changes the PR, keep PRs in review as
short as possible
- Our Short Guide for PRs:
[Link](https://docs.google.com/document/d/1T93KUtdvONq43vfTfUt8l92uo4e4SEEvFbIEKOxGr44/edit?tab=t.0)
- Remember the following Communication Standards - use comment prefixes
for clarity:
  * **blocking**: Must be addressed before approval.
  * **follow-up**: Can be addressed in a later PR or ticket.
  * **q**: Clarifying question.
  * **nit**: Non-blocking suggestions.
  * **note**: Side-note, non-actionable. Example: Praise
  * --> no prefix is considered a question