CLOUDP-306450: Simplify multi-cluster X509 test with note to revisit (#4251)
# Summary

Opening this in place of
https://github.com/10gen/ops-manager-kubernetes/pull/4227 .
Additional investigation with the agent team is required to understand
the race condition when enabling cluster authentication. Opened
https://jira.mongodb.org/browse/CLOUDP-311366 for this investigation.
This patch also removes the assertions for certificate rotation as they
were incorrectly waiting for the resource to leave `Running` state. The
assertion was only accidentally correct, because the resource would
occasionally transition to `Failed` state when waiting for the agents to
be ready would time-out.

Re-enabling certificate rotation check should use a different mechanism
for checking certificates were rotated and consumed correctly (for
example by checking automation config + agents goal state).

## Proof of Work

(Hopefully passing) test here:
https://spruce.mongodb.com/task/ops_manager_kubernetes_e2e_multi_cluster_kind_e2e_multi_cluster_tls_with_x509_patch_aadda155316f40bada5be3c13c9674e143be62f6_67f3958c729c090007eb2875_25_04_07_09_06_22/logs?execution=0

## Checklist
- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [x] Have you checked for release_note changes?

## Reminder (Please remove this when merging)
- Please try to Approve or Reject Changes the PR, keep PRs in review as
short as possible
- Our Short Guide for PRs:
[Link](https://docs.google.com/document/d/1T93KUtdvONq43vfTfUt8l92uo4e4SEEvFbIEKOxGr44/edit?tab=t.0)
- Remember the following Communication Standards - use comment prefixes
for clarity:
  * **blocking**: Must be addressed before approval.
  * **follow-up**: Can be addressed in a later PR or ticket.
  * **q**: Clarifying question.
  * **nit**: Non-blocking suggestions.
  * **note**: Side-note, non-actionable. Example: Praise
  * --> no prefix is considered a question