Fixing flaky tests (#4252)
# Summary

Fixing some flaky tests:
* `e2e_multi_cluster_backup_restore_no_mesh` it was failing because when
we were asserting that the data gets restored, the deployment was in the
process of restoring and it might actually have a re-election. Now we
wait for the resource to be reconciled properly
* `e2e_multi_cluster_replica_set` was failing because immediately after
the resource reports as Running, the pods become unready for a short
amount of time. Added retries to fix it
* `e2e_om_ops_manager_backup_sharded_cluster` was tested locally and
noticed "slow queries" because it was maxing out the 64GB of RAM.
Upgraded the variant to xlarge with 128GB

## Proof of Work

`e2e_multi_cluster_backup_restore_no_mesh` 5 consecutive passes

https://spruce.mongodb.com/task/ops_manager_kubernetes_e2e_multi_cluster_kind_e2e_multi_cluster_backup_restore_no_mesh_patch_2f2bd8a06c22aa8e47e6e77513f7e282a975a3d9_67f4d9f3ee9ccf00078bcea9_25_04_08_08_10_28/logs?execution=2&sortBy=STATUS&sortDir=ASC

`e2e_multi_cluster_replica_set` 5 consecutive passes

https://spruce.mongodb.com/version/67f5310b1c22c100072e5da7/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC

`e2e_om_ops_manager_backup_sharded_cluster` 5 consecutive passes

https://spruce.mongodb.com/version/67f52c07ea9b910007b4b9af/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC

## Checklist
- [ ] Have you linked a jira ticket and/or is the ticket in the title?
- [ ] Have you checked whether your jira ticket required DOCSP changes?
- [ ] Have you checked for release_note changes?

## Reminder (Please remove this when merging)
- Please try to Approve or Reject Changes the PR, keep PRs in review as
short as possible
- Our Short Guide for PRs:
[Link](REDACTED)
- Remember the following Communication Standards - use comment prefixes
for clarity:
  * **blocking**: Must be addressed before approval.
  * **follow-up**: Can be addressed in a later PR or ticket.
  * **q**: Clarifying question.
  * **nit**: Non-blocking suggestions.
  * **note**: Side-note, non-actionable. Example: Praise
  * --> no prefix is considered a question