CLOUDP-306289: Agent version without suffix in non-static AppDB (#4277)
# Summary

To use the agent version without suffix, I decided to partially
duplicate the method for getting the agent version since the general
function is used specifically in static container cases and it wouldn't
make sense to add another argument for the legacy case which will
eventually be removed.
The benefit of executing this fix is that customers that mirror their
images and don't use static containers would not need to mirror multiple
versions of the agent for one version of OpsManager.

## Proof of Work

Updated unit tests and e2e tests still pass.

## Checklist
- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [x] Have you checked for release_note changes?

## Reminder (Please remove this when merging)
- Please try to Approve or Reject Changes the PR, keep PRs in review as
short as possible
- Our Short Guide for PRs:
[Link](https://docs.google.com/document/d/1T93KUtdvONq43vfTfUt8l92uo4e4SEEvFbIEKOxGr44/edit?tab=t.0)
- Remember the following Communication Standards - use comment prefixes
for clarity:
  * **blocking**: Must be addressed before approval.
  * **follow-up**: Can be addressed in a later PR or ticket.
  * **q**: Clarifying question.
  * **nit**: Non-blocking suggestions.
  * **note**: Side-note, non-actionable. Example: Praise 
  * --> no prefix is considered a question