CLOUDP-310766: Set operator version to 1.0.0 (#4259)
This is required so that we can handle MEKO -> MCK upgrade correctly for
OLM as a lot of build scripts depend on the version set in
`release.json`.

This also disable smoke tests for 1.0.0 as they rely on previously
released artifacts which we do not have yet. We will re-enable after 1.0.0