CLOUDP-293808: Add hostnames to multi-cluster sharded connection string (#4263)
# Summary
This PR addresses
[CLOUDP-293808](https://jira.mongodb.org/browse/CLOUDP-293808) by adding
support for including hostnames in the connection string for
multi-cluster sharded MongoDB deployments.

**Key Changes:**

Updated the getMongoDBConnectionBuilder function to dynamically fetch
and include mongos hostnames in the connection string for sharded
clusters.

Created a dedicated helper function, getShardedClusterHostnames, to
retrieve mongos hostnames specifically for sharded clusters. This
isolates the logic for hostname fetching and improves code readability
and maintainability.

Added a new MongoDBConnectionStringBuilder which can accepts hostnames
for sharded clusters to include in the connection string.


## Proof of Work
Unit test is passing

## Checklist
- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [ ] Have you checked for release_note changes?