meko/fix-missing-om-image (#14)
# Summary

OM 7.0.12 was missing from ECR because:
1. We only support the last 3 minor versions: 7.0.13, 7.0.14, 7.0.15 so
we didn't rebuild 7.0.12 for some time
2. We had a bug in our cleanup scripts where it deleted the shas of
images instead of tags. Now it will delete the tag of an image if that
exists, or the sha if it is an untagged image.

## Proof of Work

Can't properly test the cleanup script since the 7.0.12 image was pushed
today, so need to wait 24 hours. Also can't dry-run, since the change
fixes the actual deletion.
This is the documentation for the `batch-delete-images`:
https://docs.aws.amazon.com/cli/latest/reference/ecr/batch-delete-image.html

## Checklist
- [ ] Have you linked a jira ticket and/or is the ticket in the title?
- [ ] Have you checked whether your jira ticket required DOCSP changes?
- [ ] Have you checked for release_note changes?

---------

Co-authored-by: Lucian Tosa <lucian.tosa@mongodb.com>