CLOUDP-313599 - fix todos and Documentation cleanup (#29)
# Summary


This pull request includes various changes across multiple files,
focusing on cleanup, documentation updates, and dependency management.
The most significant updates include removing outdated or redundant
documentation, updating license references, and adding a new dependency
to the project.

### Documentation and License Cleanup:
* Removed the full Apache 2.0 license text from
`mongodb-community-operator/APACHE2`, likely because it is no longer
required in the repository.
* Deleted the `mongodb-community-operator/CODE_OF_CONDUCT.md`, which
referenced the MongoDB Code of Conduct.
* Updated `mongodb-community-operator/LICENSE.md` by removing references
to the MongoDB Agent binary and the Apache license file.

### Codebase Comments and TODO Updates:
* Updated a TODO comment in `mongodb_community.py` to reference a
specific ticket (`CLOUDP-307051`) for missing "observedGeneration" in
the MongoDBCommunity status.
* Updated/Removed a bunch of TODOs which don't apply anymore

## Proof of Work

- tests are green 

## Checklist
- [ ] Have you linked a jira ticket and/or is the ticket in the title?
- [ ] Have you checked whether your jira ticket required DOCSP changes?
- [ ] Have you checked for release_note changes?