Tag latest readiness and versionhook when building on master (#57)
# Summary

Readiness probe and version hook images are not tagged with "latest"
when built on master. It fills the gap after merging repositories. All
other images are built with latest allowing to use "latest" tag for most
auxiliary images when testing locally.

## Proof of Work

If green, then the change is safe. Then we will verify it on master as
it's difficult to recreate master merge conditions in patch build.

## Checklist
- [ ] Have you linked a jira ticket and/or is the ticket in the title?
- [ ] Have you checked whether your jira ticket required DOCSP changes?
- [x] Have you checked for release_note changes?