CLOUDP-309402: MEKO MCK migration + rename enterprise resources (#18)
# Summary

## Short Summary

This PR contains 2 main set of changes:

- A new upgrade test, from MEKO to MCK. It runs in single and multi
cluster mode. The test deploys a replicaset, and perform the upgrade
steps recommended for migration.
- A rename of most components created by the operator, such as service
accounts, to make the upgrade possible without conflict between old and
new resources. When the MCK helm chart is installed, new service
accounts are created and the operator updates the stateful sets so that
the pods use the new ones.

## List of changes

### Renaming
- Introduced many variables for resource names.
- Service Accounts, RBAC Resources, Services
- Operator name, deployment name
- Selectors and labels stay the same (`mongodb-enterprise-operator`)
because for some resources it is impossible to change them without
deleting it.
- Some components kept their old names such as config map and secrets
for multi-cluster.

### Tests
- New helper functions to install the new MCK operator, to handle the
two deployments.
- Updated all existing upgrade tests to handle both deployments running
concurrently (scale down first operator before installing the second
one). This means that our existing tests are all testing a MEKO to MCK
upgrade, and some specific ones are testing other paths, such as the
sharded cluster upgrade which is doing 1.27 -> MCK.
- New MEKO to MCK dedicated test, working in both single and multi
cluster setting. Deploys a replica set and ensure the migration happens
correctly.

## Note

It is hard to ensure the renaming is completely exhaustive. The upgrade
tests ensure we renamed necessary resources to ensure a smooth upgrade
with no conflicts. Follow up PRs are/will be opened to rename smaller
things like file names.

There is still one e2e test failing that we are currently investigating,
but we preferred to open the PR early for review.

## Proof of Work

E2E test

## Checklist
- [X] Have you linked a jira ticket and/or is the ticket in the title?
- [X] Have you checked whether your jira ticket required DOCSP changes?

---------

Co-authored-by: Yavor Georgiev <yavor.georgiev@mongodb.com>
Co-authored-by: Nam Nguyen <nam.nguyen@mongodb.com>