CLOUDP-315085: Rename member list cm (#66)
# Summary

## Why

The multi cluster member-list configmap is created by both the helm
chart and kubectl tool.
It should not be created by helm. E.g one drawback is that the configmap
can be removed upon release uninstallation, which leads to a panic in
the currently running operator.
This is also confusing and it should be the responsibility of a single
tool to create that map.

The name of that configmap is 
mongodb-enterprise-operator-member-list and we should take the
opportunity to rename it to mongodb-kubernetes-operator-member-list, to
stay consistent with our new naming conventions.

Customers deploying multi cluster resources must run the kubectl plugin
when upgrading, so we can safely rename it.

## How

The config map name was mostly defined by constants so it's a small
renaming.

## Proof of Work

E2E tests pass (including upgrade tests)

---------

Co-authored-by: Mircea Cosbuc <mircea.cosbuc@mongodb.com>