Fix flakiness of e2e_appdb_tls_operator_upgrade_v1_32_to_mck (#100)
# Summary

This fix should at least reduce the flakiness of the mentioned test. The
flakiness comes from waiting for database measurements of the "admin"
db. Measurements from this db are rare, especially with no operations
happening. Therefore, we now wait for host measurements (cpu, memory,
etc.) which are constant regardless of db operations.

## Proof of Work

There are no failures due to waiting for db measurements now.

https://spruce.mongodb.com/task/mongodb_kubernetes_e2e_multi_cluster_om_appdb_e2e_appdb_tls_operator_upgrade_v1_32_to_mck_patch_6a1393da1c85ce7f02cc31986bc547bfd7adf03f_682b1e248ca26800084d419a_25_05_19_12_03_57/tests?execution=4&sortBy=STATUS&sortDir=ASC

## Checklist
- [ ] Have you linked a jira ticket and/or is the ticket in the title?
- [ ] Have you checked whether your jira ticket required DOCSP changes?
- [ ] Have you checked for release_note changes?