don't rely on version_id magic (#233)
# Summary

This pull request updates the tagging scheme in the `inventory.yaml`
file to replace the `version_id` parameter with the `version` parameter
for consistency and clarity across all image tags. The changes affect
multiple sections of the file, ensuring uniformity in how image tags are
generated.

`version_id` is a evergreen magic env which sonar magically picks up and
puts into inventory. Therefore we had those weird tags. This magic is
not required anymore - since we pass `version` on every image creation
ourselves:
https://github.com/mongodb/mongodb-kubernetes/blob/7566a394d41642b1de800ff653f9eb1dfcb92f84/pipeline.py#L265-L278.
Additionally, operator was the only leftover where we still used that
version_id


## Proof of Work
- passing ci image build

## Checklist
- [ ] Have you linked a jira ticket and/or is the ticket in the title?
- [ ] Have you checked whether your jira ticket required DOCSP changes?
- [ ] Have you checked for release_note changes?

## Reminder (Please remove this when merging)
- Please try to Approve or Reject Changes the PR, keep PRs in review as
short as possible
- Our Short Guide for PRs:
[Link](https://docs.google.com/document/d/1T93KUtdvONq43vfTfUt8l92uo4e4SEEvFbIEKOxGr44/edit?tab=t.0)
- Remember the following Communication Standards - use comment prefixes
for clarity:
  * **blocking**: Must be addressed before approval.
  * **follow-up**: Can be addressed in a later PR or ticket.
  * **q**: Clarifying question.
  * **nit**: Non-blocking suggestions.
  * **note**: Side-note, non-actionable. Example: Praise 
  * --> no prefix is considered a question