Add `OPERATOR_NAME` env var in `.generated` files for operator
As part of one of the recent changes https://github.com/mongodb/mongodb-kubernetes/pull/222
we added a new env var OPERATOR_NAME. After switching the context this env var was correctly
populated in the generated files `context.env` and `context.export.env` but it was not being
populated in the other generated files (`context.operator.env` and `context.operator.export.env`).
This commit fixes that.