Move the kubectl-mongodb plugin to cmd dir from public/tools/multicluster
This commit moves the source code of mongodb kubectl plugin from `public/tools/multicluster`
to `cmd` directory. This is to align with how go code that ships a CLI is structured.

The command/subcommand building logic is moved to the `cmd` directory and and rest of the logic

The command/subcommand building logic is moved to the `cmd` directory and and rest of the logic
is moved to the new package `pkg/kubectl-mongodb`.

This commit also moves the `.goreleaser.yaml` to the root of the project because kubectl plugin
is now at the root of the project.