Change the way kubectl-mongodb plugin is built by evergreen function
The evergreen function `build_multi_cluster_binary` is used to build the `kubectl-mongodb` plugin. And this function
gets used in the main `.evergreen.yml` file to build the `kubectl-mongodb` binary.
Since we moved the location of the source files for `kubectl-mongodb` plugin as part of [PR](https://github.com/mongodb/mongodb-kubernetes/pull/271),
this PR makes respective changes in the evergreen function so that it can properly build the `kubectl-mongodb` plugin from
the new location.