Fix OLM install modes
Currently with OLM we claim to support `OwnNamespace`, `SingleNamespace`,
and `AllNamespaces` install modes.

However, only `OwnNamespace` works properly when deploying OpsManager.
With other install modes, the operator is able to reconcile resources
and create `StatefulSets` for the workloads in OpsManager,
but pods never become healthy because only the operator's own namespace
has the required `ServiceAccount`.

With this change, we make the relevant `ServiceAccount`
managed by OLM. This means the `ServiceAccount` and relevant roles
will be created in all required namespaces.