Move the Reminder section of PR template to PR-Etiquette document
The reminder section of the PR template is supposed to be removed before merging the PR
but we forget removing that most of the time and because of that we see that in the commit
logs in master.
And it also makes sense for that to be in the PR Etiquette documents. That's why this commit
removes that section from PR template.