Move the Reminder section of PR template to PR-Etiquette document (#279)
# Summary

The reminder section of the PR template is supposed to be removed before
merging the PR but we forget removing that most of the time and because
of that we see that in the commit logs in master.
And it also makes sense for that to be in the PR Etiquette document.
That's why this PR removes that section from PR template and moves it to
the PR Etiquette
[document](https://docs.google.com/document/d/1T93KUtdvONq43vfTfUt8l92uo4e4SEEvFbIEKOxGr44/edit?tab=t.0#heading=h.f26xqmssh372).

## Proof of Work

NA

## Checklist
- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [x] Have you checked for release_note changes?