Don't expect result.suite in `pkg/kubectl-mongodb` package
Since we run the unit tests that are `pkg/kubectl-mongodb` from root
of the repo. The result.suite file will not be generated in the `pkg/kuebctl-mongodb`,
instead it will be there in the result.suite file that gets generated at the root of
the repo.
That's why we don't have to upload the file `result.suite` file from `pkg/kuebctl-mongodb`.