[CLOUDP-332196] Move `kubectl-mongodb` plugin to `cmd` dir for atomic releases (#271)
# Summary

In the efforts towards atomic releases this PR tries to move the
`kubectl-mongodb` plugin code from the directory
`public/tools/multicluster` to the root of the project `cmd/`. This is
to align the project with how other Go based projects are structured.
Apart from that this PR also moves the `.goreleaser.yaml` for the
`kubectl-mongodb` plugin from
`public/tools/multicluster/.goreleaser.yaml` directory to the root of
the repo. This is also done in order to align how `.goreleaser.yaml` is
kept in other Go based opensource projects.
This PR also moves the evergreen file that releases `kubectl-mongodb`
plugin from the location `public/.evergreen.yml` to the root of the repo
and is renamed to `.evergreen-kubectlplugin.yml`. This is to make sure
the evergreen files are kept at the same location, i.e. the root of the
repo.

### Note on LICENSE-THIRD-PARTY

The licenses that are used by the third party go libraries that we use
in our kubectl plugin, are documented in public docs at [this
link](https://www.mongodb.com/docs/mongocli/current/third-party-licenses/).
When we generated the file (`LICENSE-THIRD-PARTY`) after moving the
kubectl-mongodb code to `cmd/` directory, we can only see a new package
added, that is `sigs.k8s.io/yaml/goyaml` this seems to be part of the
module `sigs.k8s.io/yaml` which was already part of the
`LICENSE-THIRD-PARTY`, that's why we don't need to re-publish the docs
for third party licenses.

To generate the `LICENSE-THIRD-PARTY`, `make precommit` can be run after
setting env var `export MDB_UPDATE_LICENSES=true`.

This PR also fixes https://jira.mongodb.org/browse/CLOUDP-332234

## Proof of Work

Run goreleaser command locally and verify the binaries that are
generated in the `dist` directory that they output the correct
subcommands.

```
~/work/opensource/mongodb-kubernetes/dist/kubectl-mongodb_darwin_arm64_v8.0 (mongodb-kubectlplugin-refactor*) » ./kubectl-mongodb                             
This application is a tool to simplify maintenance tasks
of MongoDB resources in your kubernetes cluster.

Build: 1cd782d97de76ed25d6170a1d0aaa50382022c08, 2025-07-17T12:10:32Z

Usage:
  kubectl-mongodb [command]

Available Commands:
  completion   Generate the autocompletion script for the specified shell
  debug        Downloads all resources required for debugging and stores them into the disk
  help         Help about any command
  multicluster Manage MongoDB multicluster environments on k8s

Flags:
  -h, --help   help for kubectl-mongodb

Use "kubectl-mongodb [command] --help" for more information about a command.
```

**Testing via evergreen:**
Since we are experiencing [some
issues](https://mongodb.slack.com/archives/CGLP6R2PQ/p1752765888938289)
while testing the entire release process I am planning to merge this PR
to master and then would test the release from master.


Jira: https://jira.mongodb.org/browse/CLOUDP-332196

## Checklist
- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [x] Have you checked for release_note changes?