[CLOUDP-332196] Remove dir `public/tools` after moving kubectl plugin to `cmd/kubectl-mongodb` (#281)
# Summary

Since we moved everything related to the kubectl plugin to the dir
`cmd/kubectl-mongodb`, this commit deletes the dir `public/tools`
because it doesn't really have any valuable files now.

## Proof of Work

Run the test using the same command (`make golang-tests-race`) that is
used to run the unit tests on CI and make sure that the tests for
kubectl-plugin packages, that are in `pkg/kubectl-mongodb`, are run.

```
ok  	github.com/mongodb/mongodb-kubernetes/pkg/kubectl-mongodb/common	9.709s	coverage: 54.1% of statements
=== RUN   TestNoOpAnonymizer_AnonymizeSecret
--- PASS: TestNoOpAnonymizer_AnonymizeSecret (0.00s)
=== RUN   TestSensitiveDataAnonymizer_AnonymizeSecret
--- PASS: TestSensitiveDataAnonymizer_AnonymizeSecret (0.00s)
=== RUN   TestCollectors
--- PASS: TestCollectors (0.00s)
=== RUN   TestWriteToFile
--- PASS: TestWriteToFile (0.00s)
=== RUN   TestCleanContext
--- PASS: TestCleanContext (0.00s)
PASS
coverage: 45.8% of statements
ok  	github.com/mongodb/mongodb-kubernetes/pkg/kubectl-mongodb/debug	2.144s	coverage: 45.8% of statements
=== RUN   TestGetRsNamefromMultiStsName
--- PASS: TestGetRsNamefromMultiStsName (0.00s)
=== RUN   TestGetRsNamefromMultiStsNamePanic
--- PASS: TestGetRsNamefromMultiStsNamePanic (0.00s)
=== RUN   TestLoadKubeConfigFile
--- PASS: TestLoadKubeConfigFile (0.00s)
PASS
```

Another trivial change that this PR introduces is, now that the unit
tests related to the kubectl-mongodb plugin are moved to
`pkg/kubectl-mongodb` and we run them from the root of the repo. The
`result.suite` file will not be generated in the `pkg/kubectl-mongodb`,
instead it will be part of the `result.suite` file that gets generated
at the root of the repo. Which should be ok because we upload that tot
he evergreen using the evergreen function.

## Checklist
- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [x] Have you checked for release_note changes?